feat: 이슈 상세 화면 API Service 객체 추가
- 이슈 상세 화면 뷰모델에서 API 요청 로직 분리
