+++ iOS/IssueTracker/IssueTracker/IssueDetail/ViewModel/IssueDetailViewModel.swift
@@ -1,7 +1,11 @@
 
 import Foundation
 import Alamofire
+
 class IssueDetailViewModel {
+    
+    lazy var service = IssueDetailService(viewModel: self)
+    
     struct Status {
         var model = Bindable(IssueDetailModel.all()) // 여기에 snapshot 연결
     }
@@ -16,41 +20,13 @@ class IssueDetailViewModel {
     var status: Status
     //var action: Action
   
-    var iid: Int
-    
-//    init(model: [Comment]) {
-//        self.status = Status(model: Bindable(model))
-//    }
-//
-    init(iid: Int) {
+    var issueId: Int
+ 
+    init(issueId: Int) {
         self.status = Status()
-        self.iid = iid
-        callAPI()
+        self.issueId = issueId
+        service.callAPI(with: issueId)
     }
 
-    func callAPI() {
-        let url = "http://group05issuetracker.duckdns.org:49203"
-        
-        let parameters = ["issueID": String(self.iid)]
-
-        let headers: HTTPHeaders = ["Accept": "application/json","Authorization": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOjExLCJ1c2VySWQiOiJhYWExMTEiLCJwYXNzd29yZCI6IiQyYiQxMiRFQmIxMDBuckJCRTZZSDhXSWNvam5PS3d5YlYzSUQ1VHE4U2x6RWxnRWdXUm0xRU9TOVQzMiIsIm5pY2tuYW1lIjoiYWFhMTExIiwiT0F1dGgiOmZhbHNlLCJyZXNvdXJjZVNlcnZlciI6ImxvY2FsIiwiY3JlYXRlZEF0IjoiMjAyMC0xMS0wM1QwODozMzozNy4wMDBaIiwidXBkYXRlZEF0IjoiMjAyMC0xMS0wM1QwODozMzozNy4wMDBaIiwiaWF0IjoxNjA0NDgxMTIzfQ.p4nu83YEYOTD5NB7XKkqJTyAlgKDu6EPVgLMQqHQ_RY"]
-        
-        AF.request(url + "/api/issue/\(self.iid)", method: .get, parameters: parameters, headers: headers).responseJSON { (response) in
-            switch response.result {
-            case .success(let result):
-                do {
-                    print(result)
-                    let resultData = try JSONSerialization.data(withJSONObject: result, options: .prettyPrinted)
-                    let decodedData = try JSONDecoder().decode(IssueDetailModel.self, from: resultData)
-                    self.status.model.value = decodedData
-                } catch {
-                    print(error)
-                }
-            case .failure(let error):
-                print(error)
-            }
-        }
-    }
-    
     
 }
